Reorganize repository code
The files in the lib and spec directories are shuffled around and
renamed. Now, the application can be loaded with a single `require
'psulib_traject'` statement and the structure of the spec directory maps
to the structure of the lib directory. This makes it easier when to
switch from a class to its given unit test. A lot of editors can find
them automatically that way.

Tests that have no associated class are moved to spec/integration
because they're testing the entire system instead of a specific class.
This is like the unit versus "feature" test distinction in Rails.

Lastly, the config file that runs the whole show is moved to `config/`
as a means of keeping the application code separate from the executable
portion of the code.
