+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude`
-# on 2021-06-10 01:07:07 UTC using RuboCop version 0.93.1.
+# on 2021-06-10 15:21:05 UTC using RuboCop version 0.93.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -19,79 +19,73 @@ Metrics/BlockLength:
 # Configuration parameters: CountComments, Max, CountAsOne.
 Metrics/ClassLength:
   Exclude:
-    - 'lib/marc_format_processor.rb'
-    - 'lib/marc_media_type_processor.rb'
+    - 'lib/psulib_traject/processors/format.rb'
+    - 'lib/psulib_traject/processors/media_type.rb'
 
 # Offense count: 1
 # Configuration parameters: IgnoredMethods, Max.
 Metrics/CyclomaticComplexity:
   Exclude:
-    - 'lib/marc_media_type_processor.rb'
+    - 'lib/psulib_traject/processors/media_type.rb'
 
 # Offense count: 1
 # Configuration parameters: IgnoredMethods, Max.
 Metrics/PerceivedComplexity:
   Exclude:
-    - 'lib/traject/readers/marc_combining_reader.rb'
+    - 'lib/psulib_traject/marc_combining_reader.rb'
 
 # Offense count: 2
 # Configuration parameters: MinSize.
 Performance/CollectionLiteralInLoop:
   Exclude:
-    - 'lib/traject/readers/marc_combining_reader.rb'
+    - 'lib/psulib_traject/marc_combining_reader.rb'
 
 # Offense count: 12
 # Configuration parameters: Prefixes.
 # Prefixes: when, with, without
 RSpec/ContextWording:
   Exclude:
-    - 'spec/lib/traject/access_facet_spec.rb'
-    - 'spec/lib/traject/macros/custom_spec.rb'
-    - 'spec/lib/traject/psulib_config_spec.rb'
+    - 'spec/integration/access_facet_spec.rb'
+    - 'spec/integration/config_spec.rb'
+    - 'spec/integration/macros_spec.rb'
 
-# Offense count: 9
+# Offense count: 2
 # Configuration parameters: IgnoredMetadata.
 RSpec/DescribeClass:
   Exclude:
-    - 'spec/lib/traject/access_facet_spec.rb'
-    - 'spec/lib/traject/audience_spec.rb'
-    - 'spec/lib/traject/bound_with_spec.rb'
-    - 'spec/lib/traject/macros/custom_spec.rb'
-    - 'spec/lib/traject/media_types_spec.rb'
-    - 'spec/lib/traject/psulib_config_spec.rb'
-    - 'spec/lib/traject/pub_date_spec.rb'
-    - 'spec/lib/traject/subjects_spec.rb'
-    - 'spec/lib/traject/title_spec.rb'
+    - 'spec/integration/access_facet_spec.rb'
+    - 'spec/integration/bound_with_spec.rb'
 
 # Offense count: 9
 # Configuration parameters: AllowSubject, Max.
 RSpec/MultipleMemoizedHelpers:
   Exclude:
-    - 'spec/lib/traject/macros/custom_spec.rb'
-    - 'spec/lib/traject/title_spec.rb'
+    - 'spec/integration/macros_spec.rb'
+    - 'spec/integration/title_spec.rb'
 
 # Offense count: 3
 Style/MixinUsage:
   Exclude:
-    - 'lib/traject/psulib_config.rb'
+    - 'config/traject.rb'
 
-# Offense count: 59
+# Offense count: 60
 # Cop supports --auto-correct.
 # Configuration parameters: AutoCorrect, Max, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Layout/LineLength:
   Exclude:
     - 'config/schedule.rb'
-    - 'lib/marc_format_processor.rb'
-    - 'lib/marc_media_type_processor.rb'
-    - 'lib/traject/hathi_overlap_reducer.rb'
-    - 'lib/traject/macros/custom.rb'
-    - 'lib/traject/psulib_config.rb'
+    - 'config/traject.rb'
+    - 'lib/psulib_traject/hathi_overlap_reducer.rb'
+    - 'lib/psulib_traject/macros.rb'
+    - 'lib/psulib_traject/processors/format.rb'
+    - 'lib/psulib_traject/processors/media_type.rb'
+    - 'lib/psulib_traject/processors/pub_date.rb'
+    - 'spec/integration/bound_with_spec.rb'
+    - 'spec/integration/config_spec.rb'
+    - 'spec/integration/macros_spec.rb'
+    - 'spec/integration/subjects_spec.rb'
+    - 'spec/integration/title_spec.rb'
+    - 'spec/lib/psulib_traject/marc_combining_reader_spec.rb'
+    - 'spec/lib/psulib_traject/processors/pub_date_spec.rb'
     - 'spec/lib/psulib_traject/solr_manager_spec.rb'
-    - 'spec/lib/traject/bound_with_spec.rb'
-    - 'spec/lib/traject/macros/custom_spec.rb'
-    - 'spec/lib/traject/psulib_config_spec.rb'
-    - 'spec/lib/traject/pub_date_spec.rb'
-    - 'spec/lib/traject/readers/marc_combining_reader_spec.rb'
-    - 'spec/lib/traject/subjects_spec.rb'
-    - 'spec/lib/traject/title_spec.rb'
