+++ config/settings.yml
@@ -8,7 +8,7 @@ log:
   batch_size: 100_000
 solr_writer:
   commit_on_close: true
-reader_class_name: Traject::MarcCombiningReader
+reader_class_name: PsulibTraject::MarcCombiningReader
 commit_timeout: 10000
 symphony_data_path: /data/symphony_data/
 hathi_etas: false
