+++ spec/spec_helper.rb
@@ -9,13 +9,7 @@ ENV['RUBY_ENVIRONMENT'] = 'test'
 
 require 'rspec'
 require 'webmock/rspec'
-require 'traject'
-require_relative '../lib/traject/readers/marc_combining_reader'
-require_relative '../lib/marc_pub_date_processor'
-require_relative '../lib/marc_format_processor'
-require_relative '../lib/marc_media_type_processor'
-
-require 'pry'
+require 'psulib_traject'
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
@@ -40,7 +34,7 @@ end
 RSpec.configure do |config|
   config.before(:all) do
     @indexer ||= Traject::Indexer.new.tap do |indexer|
-      indexer.load_config_file('./lib/traject/psulib_config.rb')
+      indexer.load_config_file('./config/traject.rb')
     end
   end
